mycommit

+++ src/main/java/com/example/demospring/dto/ProductDTO.java
@@ -3,7 +3,7 @@ package com.example.demospring.dto;
 public class ProductDTO {
 
     private boolean inStock;
-    private int salePrice;
+    private double salePrice;
     private String description;
     private String title;
 
@@ -15,11 +15,11 @@ public class ProductDTO {
         this.inStock = inStock;
     }
 
-    public int getSalePrice() {
+    public double getSalePrice() {
         return salePrice;
     }
 
-    public void setSalePrice(int salePrice) {
+    public void setSalePrice(double salePrice) {
         this.salePrice = salePrice;
     }
